modify for 3d tracking
I think that position tracking needs to be implemented in 3d.  This pull (not checked for compile errors, missing some print statements) keeps track of position and velocity based on accumulated accelerations.  I think this sort of method was used on all kinds of ICBMS before there was GPS.  See "The Invention of Accuracy" for more.